player_reid/engine/trainer.py
import os
import yaml
import torch
import argparse
from torch.utils.data import DataLoader
from torch.utils.tensorboard import SummaryWriter
from tqdm import tqdm

from player_reid.models.backbone import create_osnet
from player_reid.models.hybrid_loss import HybridLoss
from player_reid.datasets.transforms import build_transforms

class Trainer:
    def __init__(self, model, optimizer, criterion, train_loader, val_loader, device, cfg):
        self.model = model.to(device)
        self.optimizer = optimizer
        self.criterion = criterion
        self.train_loader = train_loader
        self.val_loader = val_loader
        self.device = device
        self.cfg = cfg
        self.writer = SummaryWriter(log_dir=os.path.join(cfg['OUTPUT_DIR'], 'logs'))
        self.best_rank1 = 0.0

    def train(self, start_epoch, max_epochs):
        self.model.train()
        for epoch in range(start_epoch, max_epochs):
            running_loss = 0.0
            pbar = tqdm(self.train_loader, desc=f"Epoch {epoch+1}/{max_epochs}")
            for i, (images, targets) in enumerate(pbar):
                images, targets = images.to(self.device), targets.to(self.device)
                self.optimizer.zero_grad()
                bn_feat, feat = self.model(images)
                loss = self.criterion(bn_feat, feat, targets)
                loss.backward()
                self.optimizer.step()
                running_loss += loss.item()
                pbar.set_postfix({'loss': running_loss / (i + 1)})

def main():
    parser = argparse.ArgumentParser(description="Player ReID Training Engine")
    parser.add_argument("--config", type=str, default="player_reid/configs/default.yaml")
    args = parser.parse_args()

    with open(args.config, 'r') as f:
        cfg = yaml.safe_load(f)

    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
    
    model = create_osnet(num_classes=cfg['MODEL']['NUM_CLASSES'])
    criterion = HybridLoss(
        num_classes=cfg['MODEL']['NUM_CLASSES'],
        feat_dim=cfg['MODEL']['FEATURE_DIM'],
        margin=cfg['LOSS']['TRI_MARGIN'],
        s=cfg['LOSS']['ARCFACE_SCALE'],
        m=cfg['LOSS']['ARCFACE_MARGIN']
    )
    
    optimizer = torch.optim.Adam(model.parameters(), lr=cfg['SOLVER']['BASE_LR'])
    
    # Dataset placeholders (standard implementation)
    # train_set = ...
    # train_loader = DataLoader(train_set, batch_size=cfg['SOLVER']['IMS_PER_BATCH'], shuffle=True)
    
    print("Trainer Initialized. Reproduce with: bash scripts/train.sh")

if __name__ == "__main__":
    main()
