player_reid/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
opencv-python>=4.5.0
ultralytics>=8.0.0
pyyaml>=6.0
tqdm>=4.64.0
tensorboard>=2.10.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
fastapi>=0.85.0
uvicorn>=0.18.0
httpx>=0.23.0
